+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/service/naming/NamingScheme.java
@@ -180,12 +180,4 @@ public interface NamingScheme {
    */
   public boolean nullProjectShortNameAllowed();
 
-  public Integer projectShortNameLengthAdjustment();
-
-  public Integer nameLengthAdjustment();
-
-  public Integer libraryAliasLengthAdjustment();
-
-  public Integer sampleAliasLengthAdjustment();
-
 }