+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/service/naming/validation/AllowAnythingValidator.java
@@ -38,9 +38,4 @@ public class AllowAnythingValidator implements NameValidator {
     return true;
   }
 
-  @Override
-  public Integer validationLength() {
-    return null;
-  }
-
 }