+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/service/naming/validation/OicrProjectShortNameValidator.java
@@ -5,7 +5,7 @@ public class OicrProjectShortNameValidator extends RegexValidator {
   public static final String REGEX = "[A-Z0-9]{3,7}";
 
   public OicrProjectShortNameValidator() {
-    super("^" + REGEX + "$", false, false, 7);
+    super("^" + REGEX + "$", false, false);
 
     setValidationMessage("Short name must be 3-7 characters and include only capital letters and numbers");
   }