+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/service/naming/validation/RegexValidator.java
@@ -10,10 +10,8 @@ public abstract class RegexValidator implements NameValidator {
   private String customValidationMessage = null;
   private boolean allowDuplicates = false;
   private boolean allowNulls = false;
-  private Integer validationLength;
 
-  public RegexValidator(String regex, boolean allowDuplicates, boolean allowNulls, Integer validationLength) {
-    this.validationLength = validationLength;
+  public RegexValidator(String regex, boolean allowDuplicates, boolean allowNulls) {
     this.pattern = Pattern.compile(regex);
     this.allowDuplicates = allowDuplicates;
     this.allowNulls = allowNulls;
@@ -83,14 +81,5 @@ public abstract class RegexValidator implements NameValidator {
   protected abstract boolean nullabilityOptionEnabled();
 
   protected abstract boolean enableDuplicatesOptionEnabled();
-
-  @Override
-  public Integer validationLength() {
-    return validationLength;
-  }
-
-  public void setValidationLength(Integer validationLength) {
-    this.validationLength = validationLength;
-  }
 
 }