+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/util/DateType.java
@@ -1,5 +1,5 @@
 package uk.ac.bbsrc.tgac.miso.core.util;
 
 public enum DateType {
-  CREATE, UPDATE, RECEIVE, DISTRIBUTED
+  CREATE, UPDATE, RECEIVE, DISTRIBUTED, ENTERED
 }