+++ miso-web/src/main/webapp/scripts/form_project.js
@@ -6,7 +6,8 @@ FormTarget.project = (function($) {
   /*
    * Expected config {
    *   shortNameRequired: boolean
-   *   progressOptions: array
+   *   shortNameMofifiable: boolean
+   *   statusOptions: array
    * }
    */
 
@@ -50,7 +51,7 @@ FormTarget.project = (function($) {
         }, {
           title: 'Short Name',
           data: 'shortName',
-          type: 'text',
+          type: config.shortNameModifiable ? 'text' : 'read-only',
           required: config.shortNameRequired,
           maxLength: 255
         }, {
@@ -59,12 +60,12 @@ FormTarget.project = (function($) {
           type: 'text',
           maxLength: 255
         }, {
-          title: 'Progress',
-          data: 'progress',
+          title: 'Status',
+          data: 'status',
           type: 'dropdown',
           required: true,
           getSource: function() {
-            return config.progressOptions;
+            return config.statusOptions;
           }
         }, {
           title: 'Reference Genome',