+++ miso-web/src/main/webapp/scripts/list_run.js
@@ -120,7 +120,8 @@ ListTarget.run = {
         }];
   },
   searchTermSelector: function(searchTerms) {
-    return [searchTerms['id'], searchTerms['runstatus'], searchTerms['created'], searchTerms['changed'], searchTerms['creator'],
-        searchTerms['changedby'], searchTerms['platform'], searchTerms['index_name'], searchTerms['index_seq'], searchTerms['parameters']]
+    return [searchTerms['id'], searchTerms['runstatus'], searchTerms['created'], searchTerms['entered'], searchTerms['changed'],
+        searchTerms['creator'], searchTerms['changedby'], searchTerms['platform'], searchTerms['index_name'], searchTerms['index_seq'],
+        searchTerms['parameters']]
   }
 };