+++ sqlstore/src/main/java/uk/ac/bbsrc/tgac/miso/persistence/impl/HibernateBoxDao.java
@@ -304,7 +304,7 @@ public class HibernateBoxDao implements BoxStore, HibernatePaginatedDataSource<B
   @Override
   public String propertyForDate(Criteria criteria, DateType type) {
     switch (type) {
-    case CREATE:
+    case ENTERED:
       return "creationTime";
     case UPDATE:
       return "lastModified";