+++ sqlstore/src/main/resources/db/migration/V8525__project_status.sql
@@ -1,0 +1,1 @@
+ALTER TABLE Project CHANGE COLUMN progress status varchar(20) NOT NULL;