+++ RELEASE_NOTES.md
@@ -16,6 +16,13 @@ Changes:
   * Fixed validation for library aliquot targeted sequencing when library design code is changed
   * Changed the interface for modifying pool contents on the Edit Pool page
   * Autocomplete dropdown entries when pasting partial values into bulk tables
+  * Updated advanced search terms
+    * "created" now always refers to a date that the item was created in the lab, such as the
+      user-specified creation date for libraries or the start date for sequencing runs
+    * "entered" refers to a date that an item was entered into MISO
+  * Prevent changing project shortname if it is used for generating aliases and the project already
+    contains samples
+  * Renamed project field from progress to status
 
 Upgrade Notes: